:bug: fix sequelize not detecting .ts files
This entire file was auto-generated by sequelize-cli, but... there
was only one problem. It was initially made to only detect files
within the models directory with the extension of ".js", but I also
included `|| ".ts"` for it to also detect Typescript files.

Oh man this took me a minute to figure out lol.