:sparkles: create class that creates model associations
Instead of having the code that creates the model associations all
piled up in the server/entry file, I decided to just create a class
that does all of that instead. All you need to do is call
a single fn on it, and it works its magic. Much cleaner!